+++ src/screens/login/Modal.js
@@ -1,7 +1,6 @@
 import {
   View,
   Text,
-  StyleSheet,
   FlatList,
   TouchableOpacity,
   TextInput,
@@ -22,6 +21,7 @@ import {strings} from '../../constants/string';
 import {addUsers, addUid} from '../../redux/reducers/reducers';
 import {useDispatch, useSelector} from 'react-redux';
 import {chatStructure} from '../../utils/fireStore';
+import styles from './modalStyle';
 
 const Modal = ({callBack}) => {
   const keypadArray = [1, 2, 3, 4, 5, 6, 7, 8, 9, '+', 0, -1];
@@ -33,25 +33,15 @@ const Modal = ({callBack}) => {
   const [loaderState, setLoaderState] = useState(false);
   const [count, setCount] = useState(false);
 
-
   const onPress = item => {
-    console.log(selection);
     if (item === -1 && selection !== null) {
       let tempText = inputText?.split('');
-      console.log('temp arr', tempText);
-
-      {
-        selection.start > 0 &&
-          tempText.splice(
-            selection.start - 1,
-            selection.end - (selection.start - 1),
-          );
-      }
-
-      console.log('tempdeleted', tempText);
-
+      selection.start > 0 &&
+        tempText.splice(
+          selection.start - 1,
+          selection.end - (selection.start - 1),
+        );
       setInputText(tempText.join(''));
-
       selection.start > 0 &&
         setSelection({start: selection.start - 1, end: selection.end - 1});
     } else if (item === -1) {
@@ -61,7 +51,6 @@ const Modal = ({callBack}) => {
       tempText.splice(selection.start, 0, item);
       console.log(tempText.join(''));
       setSelection({start: selection.start + 1, end: selection.end + 1});
-
       setInputText(tempText.join(''));
     }
   };
@@ -87,7 +76,6 @@ const Modal = ({callBack}) => {
         inputText,
         otpResponse => {
           setCount(true);
-          setInputText('');
           setSelection({start: 0, end: 0});
           setConfirmOtp(otpResponse);
           setLoaderState(false);
@@ -125,8 +113,9 @@ const Modal = ({callBack}) => {
         </Text>
       </View>
       <Text style={styles.text2}>
-        {!count ? strings.WeSendConfirmCode : strings.WeSendToNumber}
+        {!count ? strings.WeSendConfirmCode :` ${strings.WeSendToNumber} ${inputText}`}
       </Text>
+
       <TextInput
         value={inputText}
         style={styles.textInputStyle}
@@ -136,6 +125,7 @@ const Modal = ({callBack}) => {
           setSelection(selection);
         }}
       />
+
       <FlatList
         data={keypadArray}
         contentContainerStyle={styles.flatlistStyle}
@@ -158,87 +148,5 @@ const Modal = ({callBack}) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  modalMainView: {
-    backgroundColor: colors.primaryWhite,
-    height: vh(600),
-    marginTop: 'auto',
-    padding: normalize(20),
-    borderTopEndRadius: 20,
-    borderTopStartRadius: 20,
-  },
-  loaderViewStyle: {
-    zIndex: 1,
-    height: vh(600),
-    justifyContent: 'center',
-  },
-  text1View: {
-    height: normalize(72),
-    width: normalize(186),
-  },
-  text1: {
-    fontSize: normalize(26),
-    fontWeight: '500',
-  },
-  text2: {
-    color: '#989898',
-    marginTop: normalize(10),
-    fontSize: normalize(12),
-  },
-  textInputStyle: {
-    // height: vh(48),
-    width: vw(320),
-    marginTop: vh(15),
-    fontSize: normalize(35),
-    textAlign: 'center',
-    letterSpacing: 2,
-  },
-  renderBtnView: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    margin: 1,
-    height: vh(60),
-    width: vw(106),
-  },
-  flatlistStyle: {
-    paddingRight: normalize(12),
-    alignItems: 'flex-end',
-    height: vh(260),
-  },
-  keypadBtn: {
-    height: normalize(30),
-    width: normalize(30),
-  },
-  keypadBtnText: {
-    fontSize: 23,
-  },
-  btnStyle: {
-    backgroundColor: colors.primaryColor,
-    height: normalize(40),
-    width: normalize(335),
-    marginBottom: vh(10),
-    borderRadius: 6,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  btnTextStyle: {
-    color: colors.primaryWhite,
-    fontWeight: '600',
-  },
-  bottomTextView: {
-    alignSelf: 'center',
-    height: vh(36),
-    width: vw(247),
-  },
-  bottomText: {
-    fontSize: normalize(12),
-    textAlign: 'left',
-  },
-  bottomHighlightText: {
-    fontSize: normalize(12),
-    fontWeight: 'bold',
-  },
-});
 
 export default Modal;
